backend/requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
flask-socketio>=5.3.6
python-socketio>=5.8.0
requests>=2.31.0
aiohttp>=3.8.5
anthropic>=0.7.7
openai>=1.33.0
google-generativeai>=0.3.0
mem0ai>=0.0.5
python-dotenv>=1.0.0
redis>=4.5.0
scrapybara>=1.0.0
websockets>=11.0.2
pydantic>=2.5.0
dataclasses-json>=0.6.0
asyncio-throttle>=1.0.2
psutil>=5.9.0
colorama>=0.4.6
rich>=13.0.0
